[Feature] #54917 - Reorganize the project and unify overlapping docs
Docs move to docs/, the analysis generator to scripts/, leaving reference/ as only
what the tests load as the golden. The plan restated ARCHITECTURE.md's facts, traps
and operator mapping, so those are links now and it keeps only the device
measurements and bring-up gates: 363 lines to 205, and no non-ASCII. Four errors
found while checking nothing was lost: 23 validated config fields not 17,
config.json 2.4 KB not 1.5 KB, an unbound p_bf16 in the MoE snippet, and a
no-weights test command that skipped the drift guard.
