+++ models/experimental/nomic_embed_text_v2_moe/README.md
@@ -11,7 +11,8 @@ decoder, no KV cache, no generation.
 
 | Phase | Issue | State |
 |---|---|---|
-| 0: architectural overview and PyTorch reference | [#54917](https://github.com/tenstorrent/tt-metal/issues/54917) | done, PR [#55503](https://github.com/tenstorrent/tt-metal/pull/55503) |
+| 0: architectural overview | [#54917](https://github.com/tenstorrent/tt-metal/issues/54917) | done, PR [#55503](https://github.com/tenstorrent/tt-metal/pull/55503) |
+| 0: PyTorch reference | [#54919](https://github.com/tenstorrent/tt-metal/issues/54919) | done, same PR |
 | 1: first working TTNN PoC | [#54918](https://github.com/tenstorrent/tt-metal/issues/54918) | not started |
 | 2: device performance | | not started |
 
@@ -24,21 +25,34 @@ marker.
 ## Layout
 
 ```
+README.md                     this file: status, layout, setup, test commands
+common.py                     pinned revisions, checkpoint resolution, metrics, test helpers
+docs/                         see Documentation below
 reference/
-  ARCHITECTURE.md            verified architecture and operator mapping
-  MODEL_ANALYSIS.md          generated: graph, operators, shapes, params, hierarchy
-  analysis.py                regenerates MODEL_ANALYSIS.md
-  IMPLEMENTATION_PLAN.md     approved multi-phase plan
-  modeling_nomic_moe.py      golden PyTorch reference
-  configuration_nomic_moe.py config that validates every baked-in assumption
-  loader.py                  checkpoint contract, generated from the config
-  pipeline.py                prefixes, pooling, Matryoshka, L2 normalize
-  hf_reference.py            containment for the transformers native-class trap
-  config.json                pinned config snapshot for the no-network tests
-common.py                    pinned revisions, checkpoint resolution, metrics, helpers
-tests/pcc/                   correctness tests
-tt/                          TTNN implementation (Phase 1)
+  modeling_nomic_moe.py       golden PyTorch reference
+  configuration_nomic_moe.py  config that validates every baked-in assumption
+  loader.py                   checkpoint contract, generated from the config
+  pipeline.py                 task prefixes, pooling, Matryoshka, L2 normalize
+  hf_reference.py             containment for the transformers native-class trap
+  config.json                 pinned config snapshot for the no-network tests
+scripts/
+  analysis.py                 regenerates docs/MODEL_ANALYSIS.md
+tests/
+  conftest.py                 session-scoped checkpoint, model and tokenizer fixtures
+  pcc/                        correctness tests
+tt/                           TTNN implementation (Phase 1)
 ```
+
+## Documentation
+
+Three documents, split by who writes them and what they are for. Each fact has one owner
+here; the summaries below link to it rather than repeating its numbers.
+
+| Document | Answers | Written |
+|---|---|---|
+| [`docs/ARCHITECTURE.md`](docs/ARCHITECTURE.md) | What the model is, what fails silently, and which TTNN op replaces which reference op | By hand. Every number measured against the pinned checkpoint; open assumptions quarantined in one section. |
+| [`docs/MODEL_ANALYSIS.md`](docs/MODEL_ANALYSIS.md) | Model graph, module hierarchy, aten operator inventory, per-module tensor shapes, parameters and memory | Generated by `scripts/analysis.py`. `tests/pcc/test_model_analysis.py` asserts the committed copy byte for byte, so it cannot drift. |
+| [`docs/IMPLEMENTATION_PLAN.md`](docs/IMPLEMENTATION_PLAN.md) | Phase scope, Blackhole device measurements, bring-up order and PCC gates | By hand. Updated only with verified results or approved changes. |
 
 ## Setup
 
@@ -62,12 +76,13 @@ Tests needing the checkpoint skip rather than fail when it is absent.
 pytest models/experimental/nomic_embed_text_v2_moe/tests/pcc/ -v
 
 # structural backbone: no network, no weights, no device, ~9 s
-pytest models/experimental/nomic_embed_text_v2_moe/tests/pcc/test_reference_modules.py -v
+pytest models/experimental/nomic_embed_text_v2_moe/tests/pcc/ -m "not needs_weights" -v
 ```
 
 | File | Covers | Needs weights |
 |---|---|---|
 | `test_reference_modules.py` | rotary, QKV layout, post-norm structure, router, experts, GELU, pooling, each with a negative control | no |
+| `test_model_analysis.py` | `docs/MODEL_ANALYSIS.md` still matches what the generator produces | no |
 | `test_checkpoint_contract.py` | 148 keys/shapes/dtypes generated from the config, absence assertions, strict load | yes |
 | `test_reference_vs_hf.py` | end-to-end and per-layer parity with upstream | yes, plus network |
 | `test_embedding_pipeline.py` | tokenizer, prefixes, model-card similarity, Matryoshka, ragged batches | yes, plus network |
@@ -77,17 +92,17 @@ Phase 0 is CPU-only. Do not set `TT_VISIBLE_DEVICES`; on a p300c it fails with
 
 ## Three things that fail silently
 
-Detail in [`reference/ARCHITECTURE.md`](reference/ARCHITECTURE.md).
+Each is measured, and each has a negative control in the test suite. Details and numbers in
+[`docs/ARCHITECTURE.md`](docs/ARCHITECTURE.md#3-what-is-easy-to-get-wrong).
 
 1. `AutoModel.from_pretrained` returns the wrong model without raising. transformers >= 5
    ships a native `nomic_bert` targeting v1.5, no MoE, registered for this `model_type`.
-   Loading this checkpoint into it discards every expert tensor and randomly initialises
-   `gate_proj`/`up_proj`. Always use `hf_reference.load_hf_model`.
-2. The MoE top-2 weights are not renormalized; they sum to ~0.70 on real text. Mixtral and
-   Switch both divide by the top-k sum. That bug scores PCC ~0.99, right at a typical gate.
-3. PCC cannot catch the shared-bias bug. The expert bias is one `[768]` vector added once
-   after the weighted sum; folding it into the loop gives a near-constant offset that PCC
-   mean-centres away (0.9999998 on real weights). Gate that class on max-abs.
+   Always load through `reference/hf_reference.load_hf_model`.
+2. The MoE top-2 weights are not renormalized. Mixtral and Switch both divide by the top-k
+   sum, so copying either by reflex produces a bug that lands right on a 0.99 PCC gate.
+3. PCC cannot catch the shared-bias bug at all. The expert bias is added once after the
+   weighted sum; folding it into the loop gives a near-constant offset that PCC mean-centres
+   away. Gate that class of bug on max-abs.
 
 ## References
