+++ models/experimental/nomic_embed_text_v2_moe/docs/MODEL_ANALYSIS.md
@@ -1,14 +1,17 @@
 # Reference Model Analysis
 
-Step 1 of the porting workflow: operators, shapes, parameters, module hierarchy and the
-model graph. Architecture facts and the TTNN operator mapping live in
-[`ARCHITECTURE.md`](ARCHITECTURE.md); this file is the mechanical inventory the port is
-built against.
+Step 1 of the porting workflow: the model graph, module hierarchy, operator inventory,
+per-module tensor shapes, and parameter and memory figures. Every number here is derived
+from the reference model rather than written by hand, which is the split between this
+document and [`ARCHITECTURE.md`](ARCHITECTURE.md): that one carries the verified
+architecture facts, the silent-failure list and the TTNN operator mapping.
 
-Generated by `reference/analysis.py`. Regenerate with:
+`tests/pcc/test_model_analysis.py` asserts this file byte for byte against a fresh
+regeneration, so it cannot drift from the model. Generated by `scripts/analysis.py`;
+regenerate with:
 
 ```bash
-python -m models.experimental.nomic_embed_text_v2_moe.reference.analysis
+python -m models.experimental.nomic_embed_text_v2_moe.scripts.analysis
 ```
 
 Sample input: `input_ids [2, 16]` with ragged padding [0, 5], synthetic weights at the real dimensions.
@@ -16,7 +19,7 @@ Sample input: `input_ids [2, 16]` with ragged padding [0, 5], synthetic weights 
 ## 1. Model graph
 
 Module-level, not autograd-level: `make_dot` on 475M parameters yields thousands of
-unreadable nodes. `reference/analysis.py --dot out.dot` emits the same graph in DOT for
+unreadable nodes. `scripts/analysis.py --dot out.dot` emits the same graph in DOT for
 rendering to PDF or PNG with `dot -Tpdf`.
 
 ```mermaid
