+++ models/experimental/nomic_embed_text_v2_moe/scripts/__init__.py
@@ -1,0 +1,2 @@
+# SPDX-FileCopyrightText: © 2026 Tenstorrent USA, Inc.
+# SPDX-License-Identifier: Apache-2.0
