+++ models/experimental/nomic_embed_text_v2_moe/scripts/analysis.py
@@ -9,8 +9,8 @@ the module hierarchy, and the model graph.
 
 Run from the repo root:
 
-    python -m models.experimental.nomic_embed_text_v2_moe.reference.analysis
-    python -m models.experimental.nomic_embed_text_v2_moe.reference.analysis --dot out.dot
+    python -m models.experimental.nomic_embed_text_v2_moe.scripts.analysis
+    python -m models.experimental.nomic_embed_text_v2_moe.scripts.analysis --dot out.dot
 
 Operators are captured with TorchDispatchMode on a real forward rather than torch.fx.
 symbolic_trace fails on this model twice over: torch.finfo() rejects a Proxy dtype in
@@ -38,7 +38,7 @@ from models.experimental.nomic_embed_text_v2_moe.reference.configuration_nomic_m
     load_vendored_config,
 )
 
-OUTPUT_PATH = Path(__file__).parent / "MODEL_ANALYSIS.md"
+OUTPUT_PATH = Path(__file__).parent.parent / "docs" / "MODEL_ANALYSIS.md"
 
 # Small but not degenerate: batch 2 with ragged padding exercises the mask path, and a
 # sequence longer than one tile keeps the shapes representative.
@@ -388,15 +388,18 @@ def build_markdown(config: NomicMoEConfig, model, log: OperatorLog, module_rows,
     lines = [
         "# Reference Model Analysis",
         "",
-        "Step 1 of the porting workflow: operators, shapes, parameters, module hierarchy and the",
-        "model graph. Architecture facts and the TTNN operator mapping live in",
-        "[`ARCHITECTURE.md`](ARCHITECTURE.md); this file is the mechanical inventory the port is",
-        "built against.",
+        "Step 1 of the porting workflow: the model graph, module hierarchy, operator inventory,",
+        "per-module tensor shapes, and parameter and memory figures. Every number here is derived",
+        "from the reference model rather than written by hand, which is the split between this",
+        "document and [`ARCHITECTURE.md`](ARCHITECTURE.md): that one carries the verified",
+        "architecture facts, the silent-failure list and the TTNN operator mapping.",
         "",
-        "Generated by `reference/analysis.py`. Regenerate with:",
+        "`tests/pcc/test_model_analysis.py` asserts this file byte for byte against a fresh",
+        "regeneration, so it cannot drift from the model. Generated by `scripts/analysis.py`;",
+        "regenerate with:",
         "",
         "```bash",
-        "python -m models.experimental.nomic_embed_text_v2_moe.reference.analysis",
+        "python -m models.experimental.nomic_embed_text_v2_moe.scripts.analysis",
         "```",
         "",
         f"Sample input: `input_ids [{SAMPLE_BATCH}, {SAMPLE_SEQLEN}]` with ragged padding "
@@ -405,7 +408,7 @@ def build_markdown(config: NomicMoEConfig, model, log: OperatorLog, module_rows,
         "## 1. Model graph",
         "",
         "Module-level, not autograd-level: `make_dot` on 475M parameters yields thousands of",
-        "unreadable nodes. `reference/analysis.py --dot out.dot` emits the same graph in DOT for",
+        "unreadable nodes. `scripts/analysis.py --dot out.dot` emits the same graph in DOT for",
         "rendering to PDF or PNG with `dot -Tpdf`.",
         "",
         "```mermaid",
@@ -595,6 +598,7 @@ def main() -> None:
     parser.add_argument("--dot", type=Path, default=None, help="also write the module graph as DOT")
     args = parser.parse_args()
 
+    args.out.parent.mkdir(parents=True, exist_ok=True)
     args.out.write_text(generate_markdown())
     print(f"wrote {args.out}")
